py/requirements.txt
selenium
pytesseract
pillow
numpy
faker
fake-useragent
pychalk